requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
albucore=0.0.23=pypi_0
albumentations=2.0.4=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.7.1=pypi_0
astunparse=1.6.3=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
cachetools=5.5.2=pypi_0
certifi=2024.7.4=pypi_0
charset-normalizer=3.3.2=pypi_0
click=7.1.2=pypi_0
coloredlogs=15.0.1=pypi_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.11=pypi_0
decorator=4.4.2=pypi_0
easydict=1.13=pypi_0
exceptiongroup=1.2.2=pypi_0
facenet-pytorch=2.6.0=pypi_0
ffmpeg-python=0.2.0=pypi_0
filelock=3.15.4=pypi_0
flatbuffers=1.12=pypi_0
fonttools=4.53.1=pypi_0
fsspec=2024.6.1=pypi_0
future=1.0.0=pypi_0
gast=0.4.0=pypi_0
google-auth=2.38.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.70.0=pypi_0
h11=0.12.0=pypi_0
h2=4.2.0=pypi_0
h5py=3.13.0=pypi_0
hpack=4.1.0=pypi_0
httpcore=0.13.7=pypi_0
httpx=0.19.0=pypi_0
huggingface-hub=0.29.1=pypi_0
humanfriendly=10.0=pypi_0
hyperframe=6.1.0=pypi_0
hyperpyyaml=1.2.2=pypi_0
idna=3.7=pypi_0
imageio=2.37.0=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
insightface=0.7.3=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
keras=2.9.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.5=pypi_0
lapx=0.5.9.post1=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=18.1.1=pypi_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
llvmlite=0.44.0=pypi_0
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.2=pypi_0
mdurl=0.1.2=pypi_0
ml-dtypes=0.4.1=pypi_0
more-itertools=10.6.0=pypi_0
moviepy=1.0.3=pypi_0
mpmath=1.3.0=pypi_0
namex=0.0.8=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.3=pypi_0
norbert=0.2.1=pypi_0
numba=0.61.0=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-cusparselt-cu12=0.6.2=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
onnx=1.17.0=pypi_0
onnxruntime=1.20.1=pypi_0
openai-whisper=20240930=pypi_0
opencv-python=4.11.0.86=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openssl=3.0.14=h5eee18b_0
opt-einsum=3.4.0=pypi_0
optree=0.14.0=pypi_0
packaging=24.1=pypi_0
pandas=1.5.3=pypi_0
pillow=10.2.0=pypi_0
pip=24.2=pypi_0
prettytable=3.13.0=pypi_0
proglog=0.1.10=pypi_0
protobuf=3.20.3=pypi_0
psutil=6.0.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.1.2=pypi_0
python=3.10.14=h955ad1f_1
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rfc3986=1.5.0=pypi_0
rich=13.9.4=pypi_0
rsa=4.9=pypi_0
ruamel-yaml=0.18.10=pypi_0
ruamel-yaml-clib=0.2.12=pypi_0
scikit-image=0.25.1=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.14.0=pypi_0
seaborn=0.13.2=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=72.1.0=pypi_0
simsimd=6.2.1=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
speechbrain=1.0.2=pypi_0
spleeter=2.4.0=pypi_0
sqlite=3.45.3=h5eee18b_0
stringzilla=3.11.3=pypi_0
sympy=1.13.1=pypi_0
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.9.3=pypi_0
tensorflow-estimator=2.9.0=pypi_0
tensorflow-io-gcs-filesystem=0.37.1=pypi_0
termcolor=2.5.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2025.1.10=pypi_0
tiktoken=0.9.0=pypi_0
tk=8.6.14=h39e8969_0
torch=2.6.0=pypi_0
torchaudio=2.6.0=pypi_0
torchvision=0.21.0=pypi_0
tqdm=4.66.5=pypi_0
triton=3.2.0=pypi_0
typer=0.3.2=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
ultralytics=8.3.32=pypi_0
ultralytics-thop=2.0.0=pypi_0
urllib3=2.2.2=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.43.0=pypi_0
wrapt=1.17.2=pypi_0
xz=5.4.6=h5eee18b_1
zlib=1.2.13=h5eee18b_1
